refactor(market): renamed event collecting methods to match the concept
